python-strategy-engine/grpc_server.py
"""
gRPC server for SignalOps Strategy Engine.
Listens for strategy evaluation requests and returns trading signals.

This server runs continuously, waiting for the Go execution engine
to request strategy evaluations via gRPC.
"""
import grpc
from concurrent import futures
import time
import logging
from typing import Dict
import sys

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


class StrategyEngineService:
    """Strategy Engine gRPC service implementation."""

    def SubmitOrder(self, request, context):
        """
        Handle order submission requests from Go execution engine.

        This method evaluates the strategy and returns a trading signal.
        """
        try:
            logger.info(f"Received order request for {request.symbol}")

            # For now, we'll return a simple response
            # In full implementation, this would trigger strategy evaluation
            return {
                'order_id': request.order_id,
                'status': 'PENDING',
                'message': f"Order received for {request.symbol}",
                'exchange_order_id': '',
                'executed_price': 0.0,
                'executed_quantity': 0.0,
                'fees': 0.0,
                'timestamp': int(time.time())
            }
        except Exception as e:
            logger.error(f"Order submission failed: {e}")
            context.set_code(grpc.StatusCode.INTERNAL)
            context.set_details(str(e))
            return {}

    def GetMarketData(self, request, context):
        """
        Get current market data and strategy evaluation for a symbol.
        """
        logger.info(f"Market data requested for {request.symbol}")
        return {
            'symbol': request.symbol,
            'price': 0.0,
            'timestamp': int(time.time())
        }

    def GetOrderStatus(self, request, context):
        """Get order status (stub implementation)."""
        logger.info(f"Order status requested for {request.order_id}")

        return {
            'order_id': request.order_id,
            'status': 'FILLED',
            'filled_quantity': 0.0,
            'average_price': 0.0,
            'timestamp': int(time.time())
        }

    def GetBalance(self, request, context):
        """Get account balance (stub implementation)."""
        logger.info(f"Balance requested for exchange: {request.exchange}")

        # Return mock balance
        return {
            'total_balance': 10000.0,
            'available_balance': 10000.0,
            'timestamp': int(time.time())
        }


def serve():
    """Start the gRPC server."""
    # For now, just run a simple HTTP server since we don't have protobuf files
    from flask import Flask, jsonify
    app = Flask(__name__)

    @app.route('/health', methods=['GET'])
    def health():
        return jsonify({'status': 'healthy', 'service': 'strategy-engine'})

    @app.route('/evaluate', methods=['POST'])
    def evaluate():
        return jsonify({
            'signal': 'HOLD',
            'confidence': 0.5,
            'timestamp': int(time.time())
        })

    logger.info("=" * 60)
    logger.info("SignalOps Strategy Engine Server")
    logger.info("=" * 60)
    logger.info("Server listening on port 5000")
    logger.info("gRPC on port 50051 (stub)")
    logger.info("Ready to receive strategy evaluation requests")
    logger.info("Press Ctrl+C to stop")
    logger.info("=" * 60)

    try:
        app.run(host='0.0.0.0', port=5001, debug=False)
    except KeyboardInterrupt:
        logger.info("Shutting down server...")
        logger.info("Server stopped")


if __name__ == '__main__':
    serve()
