python-strategy-engine/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
flask>=3.0.0

# Database & caching
psycopg2-binary>=2.9.9
redis>=5.0.1

# gRPC for inter-service communication
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=5.26.0

# Configuration management
pyyaml>=6.0.1

# Google Gemini (for LLM explanations)
google-genai>=0.3.0

# Data fetching
yfinance>=0.2.28
beautifulsoup4>=4.12.0

# Optional: ML/explainability enhancements
# scikit-learn>=1.3.0
# shap>=0.42.0
# lightgbm>=4.0.0
